support logr.WithCallDepth
Recently, logr introduced an API for informing a logr instance how
many call stack entries it has to skip to reach the actual call
site. When klog is configured to use a logr instance as backend, it
should use that feature.

If the logr instance doesn't care (i.e. does not implement the new
interface), then the additional parameter is simply ignored.

When the logr instance logs source code and WithCallDepth is not used,
it logs "klog.go" as caller, which isn't useful.