testinglogger: per-test, structured logging
This logr implementation can be used in tests to ensure that output
is associated with the currently running test. Compared to using the
global klog instance in a standard Go test that has some advantages:
- Log messages are associated with the currently running test.
  Test can run in parallel without interleaving their output.
- Log output is only printed when a test fails, unless "go test -v"
  is used.
- Because of that, the logger can print all log messages. It is
  not necessary to add klog.InitFlags to tests and then remember
  to increase the log level when invoking the test.

The motivation for hosting this logger in klog is that it shares code
with klogr and klog itself and that it may also be useful for others.