logcheck: detect if V().Enabled()
Besides being slightly inefficient, not storing the result of V()
in a variable is often also an indication for not using the right call inside
the if block:

  if klog.V(2).Enabled() {
      klog.InfoS("some log message")
  }

This records the message with v=0.

Correct is:

 if klogV := klog.V(2); klogV.Enabled() {
     klogV.InfoS("some log message")
 }

Detecting if clauses that do use this condition with a local variable but then
incorrectly use different logger inside the if block is a separate problem and
not solved yet.